# optimize parser options check