### 1. Explain what the PR does

d7fa48d0e **chore(parsers): add optionIsContainedInArgument**
c830bf91e **chore(parsers): opti optionsAreContainedInArgument**
b15d55997 **chore(parsers): impr OptionsAreContainedInArgument**


d7fa48d0e **chore(parsers): add optionIsContainedInArgument**

```
Currently all usages of the `optionsAreContainedInArgument` function
are using only one option. Then, `optionIsContainedInArgument` comes as
a simplification of the previous function.

Even though `optionsAreContainedInArgument` is not used anymore, it is
still being kept for now, since some parser might need to check for
multiple options in the future.

=== RUN   Benchmark_optionsAreContainedInArgumentWithOnlyOne
Benchmark_optionsAreContainedInArgumentWithOnlyOne
Benchmark_optionsAreContainedInArgumentWithOnlyOne-32            5000000               174.8 ns/op             0 B/op          0 allocs/op
=== RUN   Benchmark_optionIsContainedInArgument
Benchmark_optionIsContainedInArgument
Benchmark_optionIsContainedInArgument-32                         5000000               140.4 ns/op             0 B/op          0 allocs/op
PASS
```

c830bf91e **chore(parsers): opti optionsAreContainedInArgument**

```
optionsAreContainedInArgument now receives a variadic argument of type
uint64 instead of SystemFunctionArgument.

=== RUN   BenchmarkOptionsAreContainedInArgumentPrev
BenchmarkOptionsAreContainedInArgumentPrev
BenchmarkOptionsAreContainedInArgumentPrev-32          5000000               387.7 ns/op             0 B/op          0 allocs/op
=== RUN   BenchmarkOptionsAreContainedInArgument
BenchmarkOptionsAreContainedInArgument
BenchmarkOptionsAreContainedInArgument-32              5000000               152.3 ns/op             0 B/op
```

b15d55997 **chore(parsers): impr OptionsAreContainedInArgument**

```
Optimize OptionsAreContainedInArgument function to return as soon as it
detects a non-contained option.

Running tool: /home/gg/.goenv/versions/1.22.4/bin/go test -benchmem
-run=^$ -tags ebpf
-bench ^(BenchmarkOptionsAreContainedInArgumentOld|BenchmarkOptionsAreContainedInArgument)$
github.com/aquasecurity/tracee/pkg/events/parsers -benchtime=1000000x -race

goos: linux
goarch: amd64
pkg: github.com/aquasecurity/tracee/pkg/events/parsers
cpu: AMD Ryzen 9 7950X 16-Core Processor
=== RUN   BenchmarkOptionsAreContainedInArgumentOld
BenchmarkOptionsAreContainedInArgumentOld
BenchmarkOptionsAreContainedInArgumentOld-32             1000000               500.4 ns/op             0 B/op          0 allocs/op
=== RUN   BenchmarkOptionsAreContainedInArgument
BenchmarkOptionsAreContainedInArgument
BenchmarkOptionsAreContainedInArgument-32                1000000               351.3 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/aquasecurity/tracee/pkg/events/parsers       1.861s
```

### 2. Explain how to test it

<!--
Maintainer will review the code, and test the fix/feature, how to run Tracee ?
Give a full command line example and what to look for.
-->

### 3. Other comments

<!--
Links? References? Anything pointing to more context about the change.
-->