@rscampos 2024-08-23
double check the benchmarks and its aligned with yours:

impr OptionsAreContainedInArgument:
```
Benchmark_OptionsAreContainedInArgumentOld-4             5000000              1208 ns/op               0 B/op          0 allocs/op
Benchmark_OptionsAreContainedInArgument-4                5000000               730.6 ns/op             0 B/op          0 allocs/op
```

opti optionsAreContainedInArgument:
```
Benchmark_OptionsAreContainedInArgument-4                5000000               732.3 ns/op             0 B/op          0 allocs/op
Benchmark_OptionsAreContainedInArgumentuint64-4          5000000               259.9 ns/op             0 B/op          0 allocs/op
```

add optionIsContainedInArgument:
```
Benchmark_optionsAreContainedInArgumentWithOnlyOne-4	5000000	       321.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_optionIsContainedInArgument-4			5000000	       223.0 ns/op	       0 B/op	       0 allocs/op
```